feat(platform-ios): use loop-based alert detection
